Cota: Implement new UI
DO NOT MERGE: First patch set only includes base changes and
needs testing.

The goal here is to achieve a minimalistic UI. A UX that's based directly
upon feedback and more informal.

PS2: Removed FAB completely in replace of a bordered button with a colored
background. Fixed spacing issues with the header and update spacing dimens
between text lines. Also removed the fab drawables and added action text for
each update action, as well as updated the message fields to something a bit
more informal and simple

PS5: Remove the date from displaying beside the updated version. We want
to replace this with the Release ID soon. Also add a image to display
above the header to signify update illustration. Fix thed spacing between
the header and the message and display the CypherOS version when no updates
are available. Also add a progress bar to track the download progress and
refractor the percentage to display underneath

Change-Id: I581def57c7be3d0eb10d4527950c3d1c0beb8159