Cota: Version 3.6
Major updates for our OTA system to improve usability.

DeviceUtils have been updated to our correct device prop
and to include our custom version prop to display within
the app. We've also made use of our device prop to display
the users device when an update is unavailable as well as
the version number they're currently running. We've also
made use of this functionality for the update availabilty
message.

We've also removed support for CWM based recoveries as we
don't support it natively.

Todo: Update the layouts to disinclude whitespacing when launched.
Possibly inherit GLIF instead of mimmicking it. Also integrate COTA
into the system for seamless interaction.

Change-Id: I1d76ec10762a190f31f986ba9c6140c3ce366fd1