Cota: Use new Update Manager API for updates
This way we're able to dynamically update the message when an update
is available via our API. Google and other OEM's do this and it's
very useful for giving brief information about each update server
side.

How it works is when we set the current release version in our update
manager, we also set brief information about it in another field. The
information is then sent to our OTA app and displayed through textview
to inform the user. If a the version field is set to 0 then the system
will fallback to a default message, with no information given about the
update.

Change-Id: I3feebf03dfe43aaf90483a038361ea35a6611c80