Cota: Update notification and service handling
Cleans up notification handling and stores it in
the main update service to ensure notifications are
updated on each case properly. Also adds a a function
to cancel notifications on layout updates.

Device and version info will be displayed when no updates
are present and the last checked time now shows in
this case as well.

Change-Id: I83d36837b14d9d473a0ee28baefa9db4024416ed