Cota: fix layout corruption after checking
In the initial bringup we removed some refs for unofficial state
changes for unofficial builds. We decided we could handle that on
build time. There were some small refs to unofficial states in the
update service after checking for updates, causing the layout to
glitch out.

Let's completely remove any ref to unofficial state changes and replace
the fallback state to none, since they are similar in a way.

bug: 83000011
test: works
test: fixed
Change-Id: I441da6c8e09ea851c656d8ee4dfe555db1eb9a2b